+++ .github/workflows/dev_build_svace.yml
@@ -20,6 +20,7 @@ env:
   MODULES_MODULE_SOURCE: ${{ vars.DEV_MODULE_SOURCE }}
   MODULES_REGISTRY_LOGIN: ${{ vars.DEV_MODULES_REGISTRY_LOGIN }}
   MODULES_REGISTRY_PASSWORD: ${{ secrets.DEV_MODULES_REGISTRY_PASSWORD }}
+  MODULE_BUILD_PROFILE: "dev"
   SOURCE_REPO: "${{secrets.SOURCE_REPO}}"
   SOURCE_REPO_GIT: "${{secrets.SOURCE_REPO_GIT}}"
 
@@ -92,6 +93,7 @@ jobs:
           echo MODULES_MODULE_NAME=$MODULES_MODULE_NAME
           echo MODULES_MODULE_SOURCE=$MODULES_MODULE_SOURCE
           echo MODULE_EDITION=$MODULE_EDITION
+          echo MODULE_BUILD_PROFILE=$MODULE_BUILD_PROFILE
           echo MODULES_MODULE_TAG=$MODULES_MODULE_TAG
           echo DOCKER_CONFIG=$DOCKER_CONFIG
