+++ .github/workflows/dev_module_build.yml
@@ -21,6 +21,7 @@ env:
   MODULES_MODULE_SOURCE: ${{ vars.DEV_MODULE_SOURCE }}
   MODULES_REGISTRY_LOGIN: ${{ vars.DEV_MODULES_REGISTRY_LOGIN }}
   MODULES_REGISTRY_PASSWORD: ${{ secrets.DEV_MODULES_REGISTRY_PASSWORD }}
+  MODULE_BUILD_PROFILE: "dev"
   GO_VERSION: "1.25.9"
   GOLANGCI_LINT_VERSION: "2.11.1"
   SOURCE_REPO: "${{secrets.SOURCE_REPO}}"
@@ -346,6 +347,7 @@ jobs:
           echo MODULES_MODULE_SOURCE=$MODULES_MODULE_SOURCE
           echo MODULES_MODULE_TAG=$MODULES_MODULE_TAG
           echo MODULE_EDITION=$MODULE_EDITION
+          echo MODULE_BUILD_PROFILE=$MODULE_BUILD_PROFILE
           echo DEBUG_COMPONENT=$DEBUG_COMPONENT
           echo DOCKER_CONFIG=$DOCKER_CONFIG
