+++ .github/workflows/release_module_release-channels.yml
@@ -22,6 +22,7 @@ env:
   RELEASE_CHANNEL: ${{ github.event.inputs.channel }}
   MODULES_REGISTRY_LOGIN: ${{ vars.PROD_MODULES_REGISTRY_LOGIN }}
   MODULES_REGISTRY_PASSWORD: ${{ secrets.PROD_MODULES_REGISTRY_PASSWORD }}
+  MODULE_BUILD_PROFILE: "release"
   MODULES_MODULE_TAG: ${{ github.event.inputs.tag }}
   SOURCE_REPO: ${{ secrets.SOURCE_REPO }}
   SOURCE_REPO_GIT: ${{secrets.SOURCE_REPO_GIT}}
@@ -93,6 +94,7 @@ jobs:
           echo MODULES_MODULE_NAME=$MODULES_MODULE_NAME
           echo RELEASE_CHANNEL=$RELEASE_CHANNEL
           echo MODULE_EDITION=$MODULE_EDITION
+          echo MODULE_BUILD_PROFILE=$MODULE_BUILD_PROFILE
           echo DOCKER_CONFIG=$DOCKER_CONFIG
         shell: bash
 
@@ -180,6 +182,7 @@ jobs:
       - name: ECHO VAR
         run: |
           echo $MODULES_MODULE_SOURCE
+          echo $MODULE_BUILD_PROFILE
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
@@ -240,6 +243,7 @@ jobs:
       - name: ECHO VAR
         run: |
           echo $MODULES_MODULE_SOURCE
+          echo $MODULE_BUILD_PROFILE
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
@@ -305,6 +309,7 @@ jobs:
       - name: ECHO VAR
         run: |
           echo $MODULES_MODULE_SOURCE
+          echo $MODULE_BUILD_PROFILE
       - uses: actions/checkout@v4
         with:
           fetch-depth: 0
@@ -364,6 +369,7 @@ jobs:
       - name: ECHO VAR
         run: |
           echo $MODULES_MODULE_SOURCE
+          echo $MODULE_BUILD_PROFILE
       - name: Validation for tag
         run: |
           echo ${{ github.event.inputs.tag }} | grep -P '^v\d+\.\d+\.\d+'
