+++ .werf/images.yaml
@@ -31,6 +31,7 @@ dockerfile: Dockerfile
 {{-   $_ := set $ctx "SOURCE_REPO" $Root.SOURCE_REPO }}
 {{-   $_ := set $ctx "SOURCE_REPO_GIT" $Root.SOURCE_REPO_GIT }}
 {{-   $_ := set $ctx "MODULE_EDITION" $Root.MODULE_EDITION }}
+{{-   $_ := set $ctx "MODULE_BUILD_PROFILE" $Root.MODULE_BUILD_PROFILE }}
 {{-   $_ := set $ctx "DEBUG_COMPONENT" $Root.DEBUG_COMPONENT }}
 {{-   $_ := set $ctx "Firmware" $Root.Firmware }}
 {{-   $_ := set $ctx "Package" $Root.Package }}
