+++ crds/virtualmachines.yaml
@@ -1157,30 +1157,6 @@ spec:
                   type: object
                   description: Virtual machine statistics.
                   properties:
-                    phasesTransitions:
-                      description: Phase history.
-                      type: array
-                      items:
-                        type: object
-                        properties:
-                          phase:
-                            description: Virtual machine phase.
-                            type: string
-                            enum:
-                              - "Pending"
-                              - "Running"
-                              - "Degraded"
-                              - "Terminating"
-                              - "Stopped"
-                              - "Stopping"
-                              - "Starting"
-                              - "Migrating"
-                              - "Pause"
-                          timestamp:
-                            description: Time of transition to this phase.
-                            format: date-time
-                            nullable: true
-                            type: string
                     launchTimeDuration:
                       description: Details on the launch.
                       type: object
@@ -1422,10 +1398,6 @@ spec:
           jsonPath: .status.phase
           name: Phase
           type: string
-        - description: Time since the virtual machine entered the current phase.
-          jsonPath: .status.stats.phasesTransitions[-1].timestamp
-          name: PhaseAge
-          type: date
         - description: Real number of the virtual machine cores.
           jsonPath: .status.resources.cpu.cores
           name: Cores
