+++ werf-giterminism.yaml
@@ -8,6 +8,7 @@ config:
     - SOURCE_REPO
     - SOURCE_REPO_GIT
     - MODULE_EDITION
+    - MODULE_BUILD_PROFILE
     - DISTRO_PACKAGES_PROXY
     - SVACE_ENABLED
     - SVACE_ANALYZE_HOST
